pyproject.toml
[tool.poetry]
name = "mirk"
version = "0.1.1"
description = "Mirk is a vision-reasoning pipeline designed to interpret scenes only when needed."
authors = ["Dima Timofeev <dimkat@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/CuriousDima/mirk"
keywords = ["vision", "ai", "reasoning"]


[tool.poetry.dependencies]
black = "^24.10.0"
openai = "^1.12.0"
opencv-python = "^4.9.0"
python = "^3.12"         # Python 3.12 or less is required for PyTorch 2.2.1
python-dotenv = "^1.0.1"
torch = "2.2.1"
transitions = "^0.9.0"
ultralytics = "^8.1.28"

[tool.poetry.group.apple]
optional = true

[tool.poetry.group.apple.dependencies]
mlx = "^0.21.0"
huggingface_hub = "^0.24.0"
transformers = "^4.47.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
